Unified OrderDocument data class
