--- src/Data/OrderDocument/Pisa/pisaOrderDocumentDataClass.ts
@@ -1,8 +1,0 @@
-import { provideDataClass } from 'scrivito'
-import { pisaConfig } from '../../pisaClient'
-
-export function pisaOrderDocumentDataClass() {
-  return provideDataClass('OrderDocument', {
-    restApi: pisaConfig('order-document'),
-  })
-}
